+++ DreamGuard.BE.DAL/Models/Combo.cs
@@ -13,8 +13,8 @@ namespace DreamGuard.BE.DAL.Models
         public string Name { get; set; } = string.Empty;
         public string Slug { get; set; } = string.Empty;
         public int? AgeGroup { get; set; }
-        public double BasePrice { get; set; }
-        public double SalePrice { get; set; }
+        public decimal BasePrice { get; set; }
+        public decimal SalePrice { get; set; }
         public string Description { get; set; } = string.Empty;
         public string Color { get; set; } = string.Empty;
         public string Size { get; set; } = string.Empty;
