+++ DreamGuard.BE.DAL/Repositories/Implements/ComboRepository.cs
@@ -19,7 +19,7 @@ namespace DreamGuard.BE.DAL.Repositories.Implements
         }
 
         public async Task<PaginatedList<Combo>> GetAllCombosAsync(
-            int pageNumber, double? maxPrice, int? maxAgeGroup, string? color)
+            int pageNumber, decimal? maxPrice, int? maxAgeGroup, string? color)
         {
             // Only with active parent combos
             var query = _context.Combos
